feat: add `SessionDropdown` component
